@@ -10,16 +10,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^18.0.3",
+    "@angular/animations": "^19.2.14",
     "@angular/cdk": "^18.0.3",
-    "@angular/common": "^18.0.3",
-    "@angular/compiler": "^18.0.3",
-    "@angular/core": "^18.0.3",
-    "@angular/forms": "^18.0.3",
+    "@angular/common": "^19.2.14",
+    "@angular/compiler": "^19.2.14",
+    "@angular/core": "^19.2.14",
+    "@angular/forms": "^19.2.14",
     "@angular/material": "^18.0.3",
-    "@angular/platform-browser": "^18.0.3",
-    "@angular/platform-browser-dynamic": "^18.0.3",
-    "@angular/router": "^18.0.3",
+    "@angular/platform-browser": "^19.2.14",
+    "@angular/platform-browser-dynamic": "^19.2.14",
+    "@angular/router": "^19.2.14",
     "@types/d3": "^7.4.3",
     "bootstrap": "^5.3.3",
     "canvg": "^4.0.3",
@@ -31,13 +31,13 @@
     "rxjs": "~7.8.0",
     "sass": "^1.77.5",
     "tslib": "^2.6.3",
-    "zone.js": "~0.14.3"
+    "zone.js": "~0.15.1"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^18.0.4",
     "@angular/cli": "^20.0.1",
-    "@angular/compiler-cli": "^18.0.3",
-    "@angular/localize": "^18.0.3",
+    "@angular/compiler-cli": "^19.2.14",
+    "@angular/localize": "^19.2.14",
     "@types/jasmine": "~5.1.0",
     "jasmine-core": "~5.1.0",
     "karma": "~6.4.0",
