@@ -5,11 +5,10 @@ import { ConfigGridTestComponent } from './config-grid-test/config-grid-test.com
 import { NgIf } from '@angular/common';
 
 @Component({
-  selector: 'ct-root',
-  standalone: true,
-  imports: [CtMainMenuComponent, NavMainMenuComponent, ConfigGridTestComponent, NgIf],
-  templateUrl: './app.component.html',
-  styleUrl: './app.component.scss'
+    selector: 'ct-root',
+    imports: [CtMainMenuComponent, NavMainMenuComponent, ConfigGridTestComponent, NgIf],
+    templateUrl: './app.component.html',
+    styleUrl: './app.component.scss'
 })
 export class AppComponent {
   title = 'chitrac-ng';
