@@ -1,9 +1,10 @@
 import { Component, OnInit, Input } from '@angular/core';
 
 @Component({
-  selector: 'ct-data-bar',
-  templateUrl: './data-bar.component.html',
-  styleUrls: ['./data-bar.component.scss']
+    selector: 'ct-data-bar',
+    templateUrl: './data-bar.component.html',
+    styleUrls: ['./data-bar.component.scss'],
+    standalone: false
 })
 export class DataBarComponent implements OnInit {
 
