@@ -1,9 +1,10 @@
 import { Component, OnInit, Input } from '@angular/core';
 
 @Component({
-  selector: 'ct-lane-stop',
-  templateUrl: './lane-stop.component.html',
-  styleUrls: ['./lane-stop.component.scss']
+    selector: 'ct-lane-stop',
+    templateUrl: './lane-stop.component.html',
+    styleUrls: ['./lane-stop.component.scss'],
+    standalone: false
 })
 export class LaneStopComponent implements OnInit {
 
