@@ -6,11 +6,10 @@ import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatExpansionModule } from '@angular/material/expansion';
 
 @Component({
-  selector: 'app-ct-main-menu',
-  standalone: true,
-  imports: [MatIconModule, MatMenuModule, MatButtonModule, MatToolbarModule, MatExpansionModule],
-  templateUrl: './ct-main-menu.component.html',
-  styleUrl: './ct-main-menu.component.css'
+    selector: 'app-ct-main-menu',
+    imports: [MatIconModule, MatMenuModule, MatButtonModule, MatToolbarModule, MatExpansionModule],
+    templateUrl: './ct-main-menu.component.html',
+    styleUrl: './ct-main-menu.component.css'
 })
 export class CtMainMenuComponent {
 
