@@ -7,11 +7,10 @@ import { OperatorGridComponent } from '../operator-grid/operator-grid.component'
 import { OperatorConfig } from '../shared/models/operator.model';
 
 @Component({
-  selector: 'settings-operators',
-  standalone: true,
-  imports: [MatCardModule, OperatorGridComponent],
-  templateUrl: './settings-operators.component.html',
-  styleUrl: './settings-operators.component.scss'
+    selector: 'settings-operators',
+    imports: [MatCardModule, OperatorGridComponent],
+    templateUrl: './settings-operators.component.html',
+    styleUrl: './settings-operators.component.scss'
 })
 export class SettingsOperatorsComponent {
 
